+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CentOS-7 7.5.1804 x86_64, Apache 2.4, PHP-FPM 7.2, PHP memcached 3.0, Zend Opcac
 - Updates `elinks` package to elinks-0.12-0.37.pre6.el7.0.1.
 - Updates `httpd24u` packages to 2.4.38-1.
 - Updates `php72u` packages to 7.2.16-1.
+- Adds consideration for event lag into test cases for unhealthy health_status events.
 
 ### 3.1.1 - 2018-12-03